Android project rebuild
We found a way to accomplish the tab menus with swipe interaction, but to implement it, it was easier to rebuild the entire android project. Nothing was coded yet, so no loss: just make sure that your android project is set to "FoodMate" and not "Android" in the P2 folder.